chore(constants): Move mock data out to separate file
ref #34
